+++ noodles-gtf/examples/gtf_write.rs
@@ -4,13 +4,14 @@
 
 use std::io;
 
+use noodles_gff as gff;
 use noodles_gtf as gtf;
 
 fn main() -> io::Result<()> {
     let stdout = io::stdout().lock();
     let mut writer = gtf::io::Writer::new(stdout);
 
-    let record = gtf::RecordBuf::default();
+    let record = gff::feature::RecordBuf::default();
     writer.write_record(&record)?;
 
     Ok(())